+++ argilla/src/extralit/extraction/chunking.py
@@ -3,6 +3,7 @@ from os.path import join, exists
 from typing import Optional, List, Tuple, Dict, Any, Iterable
 
 import argilla as rg
+from extralit.storage.files import FileHandler, StorageType
 import pandas as pd
 from llama_index.core.schema import Document
 
@@ -21,6 +22,8 @@ def create_nodes(
     preprocessing_dataset: Optional[rg.FeedbackDataset] = None,
     response_status=['submitted'],
     exclude_llm_metadata_keys=EXCLUDE_LLM_METADATA_KEYS,
+    storage_type: StorageType=StorageType.FILE,
+    bucket_name: Optional[str]=None,
     **nougat_kwargs,
 ) -> Tuple[List[Document], List[Document]]:
     """
@@ -46,14 +49,15 @@ def create_nodes(
 
     if preprocessing_dataset is not None:
         # Load the segments from the manually annotated preprocessing dataset
-        text_segments, _, _ = create_or_load_nougat_segments(paper, **nougat_kwargs)
+        text_segments, _, _ = create_or_load_nougat_segments(paper, storage_type=storage_type, bucket_name=bucket_name, **nougat_kwargs)
         table_segments = get_paper_tables(paper, preprocessing_dataset, response_status=response_status)
     else:
-        # Load the segments from the preprocessed data
+        # Load the segments from `nougat` preprocessed data
         texts_path = join(preprocessing_path, reference, 'texts.json')
         tables_path = join(preprocessing_path, reference, 'tables.json')
-        text_segments = Segments.parse_file(texts_path) if exists(texts_path) else Segments()
-        table_segments = Segments.parse_file(tables_path) if exists(tables_path) else Segments()
+        file_handler = FileHandler(preprocessing_path, storage_type, bucket_name)
+        text_segments = Segments.parse_raw(file_handler.read_text(texts_path)) if file_handler.exists(texts_path) else Segments()
+        table_segments = Segments.parse_raw(file_handler.read_text(tables_path)) if file_handler.exists(tables_path) else Segments()
 
     extra_metadata = {'reference': reference}
     text_nodes = create_text_nodes(